FMT/FMT/requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.17
django-cors-headers==2.5.2
django-leaflet==0.24.0
django-phone-field==1.7.0
django-stubs==0.8.2
djangorestframework==3.9.2
djangorestframework-stubs==0.2.0
freetype-py==2.0.0.post6
idna==2.8
mypy==0.670
mypy-extensions==0.4.1
numpy==1.16.2
Pillow==5.4.1
psycopg2==2.7.7
pygal==2.4.0
pypng==0.0.19
pytz==2018.9
requests==2.21.0
six==1.12.0
typed-ast==1.3.1
typing-extensions==3.7.2
urllib3==1.24.1
python-dateutil==2.6.1
pandas==0.21.0
django-pandas==0.4.5
django-pdb==0.6.1
